Refactor : 통합광고현황 폴더 구조 변경
