+++ src/routes/Dashboard/AdStatus/adStatus.module.scss
@@ -1,0 +1,83 @@
+@use '/src/styles/constants/colors';
+
+.adSectionWrapper {
+  height: 100%;
+  margin-top: 100px;
+
+  .adSectionTitle {
+    margin-left: 40px;
+    font-size: 16px;
+    font-weight: 700;
+    color: colors.$GRAY800;
+  }
+
+  .boardWrapper {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    padding: 40px;
+    margin: 20px 40px 40px;
+    background-color: colors.$WHITE;
+    border-radius: 20px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 4%);
+
+    .adCardWrapper {
+      display: grid;
+      grid-template-columns: repeat(3, 1fr);
+      gap: 20px;
+
+      li {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: space-between;
+        padding: 18px 40px;
+        border: 0.5px solid colors.$GRAY50;
+        border-radius: 10px;
+
+        dt {
+          margin-bottom: 10px;
+          font-size: 12px;
+          font-weight: 500;
+          color: colors.$GRAY300;
+        }
+
+        dd {
+          font-size: 16px;
+          font-weight: 700;
+          color: colors.$GRAY800;
+        }
+
+        .changeWrapper {
+          align-self: flex-end;
+          font-size: 12px;
+          font-weight: 500;
+          color: colors.$GRAY300;
+
+          span {
+            padding-left: 5px;
+          }
+        }
+      }
+    }
+
+    .dropDownsContainer {
+      z-index: 1;
+      display: flex;
+      justify-content: space-between;
+      margin-top: 30px;
+
+      .dropDownWrapper {
+        display: flex;
+
+        :first-child {
+          margin-right: 5px;
+        }
+      }
+    }
+
+    .chartContainer {
+      z-index: 0;
+    }
+  }
+}
