--- src/routes/Dashboard/Chart/chart.module.scss
@@ -1,18 +1,0 @@
-.dropDownsContainer {
-  z-index: 1;
-  display: flex;
-  justify-content: space-between;
-  margin-top: 30px;
-
-  .dropDownWrapper {
-    display: flex;
-
-    :first-child {
-      margin-right: 5px;
-    }
-  }
-}
-
-.chartContainer {
-  z-index: 0;
-}
