+++ src/routes/Dashboard/index.tsx
@@ -1,24 +1,15 @@
 import { useQuery } from 'react-query'
-import { useState, useMount } from 'hooks'
+import { useState } from 'hooks'
 import { useRecoilState, useSetRecoilState } from 'hooks/state'
 import store from 'store'
 
-import {
-  byChannelDataResultState,
-  byChannelFetchState,
-  dailyDataResultState,
-  dailyFetchState,
-  prevDailyDataResultState,
-} from 'states/dashboard'
-import { getDailyData, getByChannelData } from 'services/ads'
+import { byChannelDataResultState, byChannelFetchState } from 'states/dashboard'
+import { getByChannelData } from 'services/ads'
 
-import AdCardContent from './AdCardContent'
+import AdStatus from './AdStatus'
 import CalendarModal from './CalendarModal/CalendarModal'
-import Chart from './Chart'
 import CurrentStatusOfMedium from './CurrentStatusOfMedium'
-
 import { DownArrow } from 'assets/svgs'
-import Loading from 'routes/_shared/Loading'
 import styles from './dashboard.module.scss'
 
 const Dashboard = () => {
@@ -26,55 +17,9 @@ const Dashboard = () => {
 
   const currentStartDate = store.get('startDate')
   const currentEndDate = store.get('endDate')
-  const prevStartDate = store.get('prevStartDate')
-  const prevEndDate = store.get('prevEndDate')
 
-  const [dailyFetch, setDailyFetch] = useRecoilState(dailyFetchState)
   const [byChannelFetch, setByChannelFetch] = useRecoilState(byChannelFetchState)
-  const setDailyData = useSetRecoilState(dailyDataResultState)
-  const setPrevDailyData = useSetRecoilState(prevDailyDataResultState)
   const setByChannelData = useSetRecoilState(byChannelDataResultState)
-
-  useMount(() => {
-    getDailyData(currentStartDate, currentEndDate).then((res) => {
-      setDailyData(res.data)
-    })
-    getDailyData(prevStartDate, prevEndDate).then((res) => {
-      setPrevDailyData(res.data)
-    })
-  })
-
-  const { isLoading: isDailyLoading } = useQuery(
-    ['getDailyData', currentStartDate, currentEndDate],
-    () => {
-      getDailyData(currentStartDate, currentEndDate).then((res) => {
-        setDailyData(res.data)
-      })
-    },
-    {
-      useErrorBoundary: true,
-      enabled: !!dailyFetch,
-      onSuccess: () => {
-        setDailyFetch(false)
-      },
-    }
-  )
-
-  const { data: prevDailyDataResult } = useQuery(
-    ['getPrevDailyData', prevStartDate, prevEndDate],
-    () => {
-      getDailyData(prevStartDate, prevEndDate).then((res) => {
-        setPrevDailyData(res.data)
-      })
-    },
-    {
-      useErrorBoundary: true,
-      enabled: !!dailyFetch,
-      onSuccess: () => {
-        setDailyFetch(false)
-      },
-    }
-  )
 
   const { isLoading: isChannelLoading } = useQuery(
     ['getByChannelData', currentStartDate, currentEndDate],
@@ -95,10 +40,6 @@ const Dashboard = () => {
     setIsModalOpen(true)
   }
 
-  if (isDailyLoading || isChannelLoading) {
-    return <Loading />
-  }
-
   return (
     <div>
       <header className={styles.header}>
@@ -116,13 +57,7 @@ const Dashboard = () => {
         </div>
       </header>
       <main className={styles.main}>
-        <div className={styles.adSectionWrapper}>
-          <h2 className={styles.adSectionTitle}>통합 광고 현황</h2>
-          <div className={styles.boardWrapper}>
-            <AdCardContent />
-            <Chart />
-          </div>
-        </div>
+        <AdStatus />
         <div className={styles.currentStatusOfMediumSectionWrapper}>
           <h2 className={styles.currentStatusOfMediumTitle}>매체 현황</h2>
           <CurrentStatusOfMedium />
